+++ R/stan.R
@@ -65,8 +65,8 @@ epinow2_rstan_model <- function(model = "estimate_infections") {
 ##'
 ##' @param model A character string indicating the model to use. One of
 ##' "estimate_infections" (default), "simulate_infections",
-##' "estimate_secondary", "simulate_secondary", "estimate_truncation" or
-##' "dist_fit".
+##' "estimate_secondary", "simulate_secondary", "estimate_truncation",
+##' "estimate_dist", or "dist_fit".
 ##' @inheritParams stan_opts
 ##' @return A stan model object (either \code{rstan::stanmodel} or
 ##'   \code{cmdstanr::CmdStanModel}, depending on the backend)
@@ -79,6 +79,7 @@ epinow2_stan_model <- function(backend = c("rstan", "cmdstanr"),
                                  "estimate_secondary",
                                  "simulate_secondary",
                                  "estimate_truncation",
+                                 "estimate_dist",
                                  "dist_fit"
                                )) {
   backend <- arg_match(backend)